Remove the `VersionValue`
